Add Fundamentals of DevOps and Software Delivery
I'm humbled and grateful that my previous book, _Terraform: Up & Running_, is already in the list of books here, so I figured it would make sense to add my latest book, _[Fundamentals of DevOps and Software Delivery](https://www.fundamentals-of-devops.com/)_ as well. I also fixed the GitHub usernames for Gruntwork and myself.